+++ tests/tests.py
@@ -5,6 +5,7 @@ import pandas as pd
 import neurokit as nk
 
 from test_ecg import TestEcg
+from test_signal import TestComplexity
 from test_miscellaneous import TestMiscellaneous
 from test_statistics import TestStatistics